Fixed index error during config backup on updates

+++ sickrage/core/helpers/__init__.py
@@ -986,17 +986,17 @@ def backupSR(backupDir, keep_latest=False):
         import glob
 
         _files = glob.glob(os.path.join(backupDir, '*.zip'))
+        if _files:
+            now = time.time()
+            newest = _files[0], now - os.path.getctime(_files[0])
+            for f in _files[1:]:
+                age = now - os.path.getctime(f)
+                if age < newest[1]:
+                    newest = f, age
+            _files.remove(newest[0])
 
-        now = time.time()
-        newest = _files[0], now - os.path.getctime(_files[0])
-        for f in _files[1:]:
-            age = now - os.path.getctime(f)
-            if age < newest[1]:
-                newest = f, age
-        _files.remove(newest[0])
-
-        for f in _files:
-            os.remove(f)
+            for f in _files:
+                os.remove(f)
 
     if keep_latest:
         _keep_latest_backup()
